datadog/requirements.txt

datadog-api-client==2.31.0

mcp==1.2.0

python-dotenv==1.0.1

loguru  ==0.7.3

icecream    ==2.1.3

fastmcp==0.4.1